Add log slider and use it in step of breathing and rainbow

+++ app/lib/bloc/effects/breathing_bloc.dart
@@ -33,7 +33,7 @@ class BreathingColorEvent extends BreathingEffectEvent {
 }
 
 class BreathingStepEvent extends BreathingEffectEvent {
-  double step;
+  int step;
 
   BreathingStepEvent(this.step);
 }
